stamp styles for horizontal verses layout
