+++ src/styles/_colors.scss
@@ -1,0 +1,3 @@
+ $light-grey: rgba(0, 0, 0, .15);
+ $white: #ffffff;
+ $dark-red: #c20000;
